+++ client/main.html
@@ -10,6 +10,6 @@
 </head>
 
 <body>
-  <div id="react-root" class="d-flex flex-column h-100"></div>
+  <div id="react-root"></div>
   <noscript>JavaScript is required to use this app</noscript>
 </body>
