+++ imports/ui/admin/components/AdminUserProfile/AdminUserProfile.jsx
@@ -88,7 +88,6 @@ const AdminUserProfile = ({ user, updateUser, removeUser }) => {
               <Form.Control
                 type="text"
                 name="firstName"
-                className="form-control"
                 placeholder="First Name"
                 defaultValue={(user && user.profile && user.profile.firstName) || ''}
                 {...register('firstName', { required: "What's your first name?" })}
@@ -104,7 +103,6 @@ const AdminUserProfile = ({ user, updateUser, removeUser }) => {
               <Form.Control
                 type="text"
                 name="lastName"
-                className="form-control"
                 placeholder="Last Name"
                 defaultValue={(user && user.profile && user.profile.lastName) || ''}
                 {...register('lastName', { required: "What's your last name?" })}
@@ -118,7 +116,6 @@ const AdminUserProfile = ({ user, updateUser, removeUser }) => {
           <Form.Control
             type="email"
             name="emailAddress"
-            className="form-control"
             placeholder="Email Address"
             defaultValue={emailAddress || ''}
             {...register('emailAddress', {
@@ -143,7 +140,6 @@ const AdminUserProfile = ({ user, updateUser, removeUser }) => {
           <Form.Control
             type="password"
             name="currentPassword"
-            className="form-control"
             placeholder="Password"
             {...register('currentPassword', {
               minLength: { value: 6, message: 'Please use at least six characters.' },
@@ -159,7 +155,6 @@ const AdminUserProfile = ({ user, updateUser, removeUser }) => {
           <Form.Control
             type="password"
             name="newPassword"
-            className="form-control"
             placeholder="Password"
             {...register('newPassword', {
               minLength: { value: 6, message: 'Please use at least six characters.' },
