+++ imports/ui/global/components/PageHeader/PageHeader.jsx
@@ -3,12 +3,10 @@ import PropTypes from 'prop-types';
 import Styles from './styles';
 
 const PageHeader = ({ title, subtitle }) => (
-  <Styles.PageHeader>
-    <Styles.PageHeaderContainer>
-      <h1>{title}</h1>
-      {subtitle && <p>{subtitle}</p>}
-    </Styles.PageHeaderContainer>
-  </Styles.PageHeader>
+  <Styles.Wrapper>
+    <h2>{title}</h2>
+    {subtitle && <p>{subtitle}</p>}
+  </Styles.Wrapper>
 );
 
 PageHeader.defaultProps = {
