+++ imports/ui/layouts/WithSidebar/WithSidebar.jsx
@@ -14,23 +14,25 @@ const WithSidebar = (props) => {
   const location = useLocation();
   const navigate = useNavigate();
 
-  // If the user should be logged in, and they are not logged in -- redirect to the login page
-  if (authRequired && !userId) {
-    return navigate(`/login${`?redirect=${location.pathname}` || ''}`);
-  }
-
-  // If this is a public page and the user is logged in -- redirect to the index page
-  if (isPublic && userId) {
-    return navigate('/');
-  }
-
-  // If this page requires a role, check that the user is in that role
-  if (authRequired && userId && reqRole) {
-    const hasRole = isInRole(allowedRoles);
-    if (!hasRole) {
-      return navigate('/');
+  useEffect(() => {
+    // If the user should be logged in, and they are not logged in -- redirect to the login page
+    if (authRequired && !userId) {
+      navigate(`/login${`?redirect=${location.pathname}` || ''}`);
     }
-  }
+
+    // If this is a public page and the user is logged in -- redirect to the index page
+    if (isPublic && userId) {
+      navigate('/');
+    }
+
+    // If this page requires a role, check that the user is in that role
+    if (authRequired && userId && reqRole) {
+      const hasRole = isInRole(allowedRoles);
+      if (!hasRole) {
+        navigate('/');
+      }
+    }
+  }, []);
 
   // Otherwise -- show the page
   return (
