+++ imports/ui/users/components/LoginForm/index.js
@@ -1,0 +1,1 @@
+export { default } from './LoginForm';
