+++ imports/ui/users/components/LoginForm/styles.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-const Login = styled.div`
+const LoginForm = styled.div`
   border: 1px solid var(--gray-lighter);
   border-radius: 3px;
   margin: 0 auto;
@@ -11,17 +11,8 @@ const Login = styled.div`
     margin-top: 0;
   }
 
-  .text-right {
-    display: block;
-  }
-
   > .row {
     margin: 0 !important;
-  }
-
-  button[type='submit'] {
-    height: 41px;
-    margin-top: 20px;
   }
 
   @media screen and (min-width: 768px) {
@@ -31,5 +22,5 @@ const Login = styled.div`
 `;
 
 export default {
-  Login,
+  LoginForm,
 };
